Rewrite memory counting to be more granular and accurate:
memory is now counted in reads and writes separately, and we now count bytes of memory transferred rather than number of memory accesses made. We do not properly count memory accessed by calls to malloc&et al nor memory access caused by function calls and stack spilling. We assume a hypothetically best case of no memory spilling, since we cannot know ahead of time how bad the generated code will be with regards to spilling. Besides, this should ultimately be on the level of noise and thus should be a fine concession to make.